fix homework2
